utils.py: add missing dtype argument to sym_normalize_matrix

+++ gcn/utils.py
@@ -40,11 +40,11 @@ def dropout_sparse(x, keep_prob, num_nonzero_elems):
     longer amounts to mere averaging of neighboring nodes). Combining these two tricks, we essentially arrive at the 
     propagation rule introduced in Kipf & Welling (ICLR 2017):
     """
-def sym_normalize_matrix(adj):
+def sym_normalize_matrix(adj, dtype=np.float32):
     adj_ = sp.coo_matrix(adj)
     rowsum = np.array(adj_.sum(1))
     degree_mat_inv_sqrt = sp.diags(np.power(rowsum, -0.5).flatten())
-    return adj_.dot(degree_mat_inv_sqrt).transpose().dot(degree_mat_inv_sqrt).tocoo()
+    return adj_.dot(degree_mat_inv_sqrt).transpose().dot(degree_mat_inv_sqrt).astype(dtype).tocoo()
 
 
 #
